+++ app/controllers/api/v1/issues_controller.rb
@@ -287,6 +287,14 @@ module Api
         return false
       end
 
+      def find_issues
+        @issues = WorkPackage.find_all_by_id(params[:id] || params[:ids])
+        raise ActiveRecord::RecordNotFound if @issues.empty?
+        @projects = @issues.collect(&:project).compact.uniq
+        @project = @projects.first if @projects.size == 1
+      rescue ActiveRecord::RecordNotFound
+        render_404
+      end
     end
   end
 end