+++ app/helpers/application_helper.rb
@@ -266,7 +266,7 @@ module ApplicationHelper
 
   def join_flash_messages(messages)
     if messages.respond_to?(:join)
-      messages.join('<br />').html_safe 
+      messages.join('<br />').html_safe
     else
       messages
     end
@@ -954,12 +954,12 @@ module ApplicationHelper
       ret += content_tag :ul do
 		    args[:collection].collect do |(s, name)|
           content_tag :li do
-            context_menu_link (name || s), work_package_bulk_update_path(:ids => args[:updated_object_ids],
-                                                                         :work_package => { db_attribute => s },
-                                                                         :back_url => args[:back_url]),
-                                                                         :method => :put,
-                                                                         :selected => args[:selected].call(s),
-                                                                         :disabled => args[:disabled].call(s)
+            context_menu_link (name || s), work_packages_bulk_path(:ids => args[:updated_object_ids],
+                                                                   :work_package => { db_attribute => s },
+                                                                   :back_url => args[:back_url]),
+                                           :method => :put,
+                                           :selected => args[:selected].call(s),
+                                           :disabled => args[:disabled].call(s)
           end
         end.join.html_safe
       end